requirements.txt
fpdf
PyPDF2
tabulate
cryptography
phonenumbers
psutil
pytube
speedtest-cli
colorama